feat(Users): Correctly throw error when libsodiuum issue
With dberror
