+++ Network/Database/Models/Users/Users.cpp
@@ -5,11 +5,12 @@
 ** Users
 */
 #include "Users.hpp"
+#include "DbError.hpp"
 
 Users::Users(Database *db) : _dbcore(db), _db(db->db), _rc(db->rc) {
     createUsersTable();
-    // if (sodium_init() < 0)
-    //     throw new DbError("Couldn't load sodium...");
+    if (sodium_init() < 0)
+        throw new DbError("Couldn't load sodium...");
 };
 
 bool Users::userIdExists(int id) {
